Address review: simplify selectItemsOfType, fix essence cap, split tests
- selectItemsOfType drops the manual per-type Selector cache: reselect
  v5's default weakMapMemoize already caches per unique (state, type)
  combination, so createSelector([...], combiner) alone is sufficient.
  It now returns plain ItemData[] rather than the generic ItemDataFor<T>[]
  narrowing; buildItemCatalog's byType facet casts at that boundary.
- Essence's max is always 6 (BASE_ESSENCE), not the attribute
  namespace's essence info — selectEssence no longer takes an
  essenceMax argument, and buildItemCatalog no longer needs the
  attribute catalog passed in at all. useRunnerSelector goes back to
  a single inline object literal.
- Split runnerSelector.test.tsx into one file per namespace, colocated
  with its catalog: attribute.catalog.test.ts, damage.catalog.test.ts,
  item.catalog.test.ts, modifiers.catalog.test.ts. runnerSelector.test.tsx
  keeps only the reactivity test, which is the hook's own concern
  rather than any one catalog's.
