+++ src/lib/stores/runner/selectors/item.catalog.ts
@@ -1,13 +1,11 @@
 import { selectById, selectEquipped } from "#/lib/stores/runner/gear/gearSlice.selectors.ts"
 import type { UUID } from "#/lib/uuidUtils.ts"
-import { AttributeKey } from "#/system/attributeKey.ts"
 import type { ArmorRating } from "#/system/gear/armorData.ts"
 import type { ItemData } from "#/system/itemData.ts"
 import type { ItemType } from "#/system/itemType.ts"
 import type { ItemDataFor } from "#/system/items/itemUtils.ts"
 import type { RunnerData } from "#/system/runnerData.ts"
 
-import type { RunnerAttributeCatalog } from "./attribute.catalog.ts"
 import type { ItemEssenceFacets } from "./item.selectors.ts"
 import { selectArmorEffective, selectArmorTotal, selectEssence, selectItemsOfType } from "./item.selectors.ts"
 
@@ -22,16 +20,16 @@ export interface RunnerItemCatalog {
   essence: ItemEssenceFacets
 }
 
-export function buildItemCatalog(state: RunnerData, attribute: RunnerAttributeCatalog): RunnerItemCatalog {
+export function buildItemCatalog(state: RunnerData): RunnerItemCatalog {
   const catalog = (id: UUID): ItemData | undefined => selectById(id)(state)
 
   return Object.assign(catalog, {
-    byType: <T extends ItemType>(type: T): ItemDataFor<T>[] => selectItemsOfType(state, type),
+    byType: <T extends ItemType>(type: T): ItemDataFor<T>[] => selectItemsOfType(state, type) as ItemDataFor<T>[],
     equipped: selectEquipped(state),
     armor: {
       total: selectArmorTotal(state),
       effective: selectArmorEffective(state),
     },
-    essence: selectEssence(state, attribute(AttributeKey.essence).info.max),
+    essence: selectEssence(state),
   })
 }
